# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateAPIKeyInputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/CreateAPIKeyInputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "name": {
            "description": "Human-readable label for this key",
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "scope": {
            "default": "read-write",
            "description": "Key scope: read (GET only) or read-write (full access)",
            "enum": [
              "read",
              "read-write"
            ],
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ListAPIKeysOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ListAPIKeysOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "keys": {
            "items": {
              "$ref": "#/components/schemas/KeyMetaResponse"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "keys"
        ],
        "type": "object"
      },
...
      "SystemStatusOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/SystemStatusOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "db": {
            "$ref": "#/components/schemas/DBStatus"
          },
          "enrichment": {
            "$ref": "#/components/schemas/EnrichmentStatus"
          },
          "nats": {
            "$ref": "#/components/schemas/NATSStatus"
          },
          "scan_jobs": {
            "$ref": "#/components/schemas/ScanJobsStatus"
          },
          "scanner": {
            "$ref": "#/components/schemas/ScannerStatus"
          }
        },
        "required": [
          "enrichment",
          "scanner",
          "nats",
          "scan_jobs",
          "db"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "ocidex_\u003ctoken\u003e",
        "description": "API key issued via POST /api/v1/auth/keys",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "description": "Session cookie obtained via GitHub OAuth (/auth/login)",
        "in": "cookie",
        "name": "ocidex_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Open Container Initiative Dex — SBOM metadata management service",
    "title": "OCIDex API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/status": {
      "get": {
        "description": "Admin-only.",
        "operationId": "get-system-status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get system status",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/auth/keys": {
      "get": {
        "operationId": "list-api-keys",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAPIKeysOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List API keys",
        "tags": [
          "Auth"
        ]
      },
      "post": {
        "operationId": "create-api-key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAPIKeyInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAPIKeyOutputBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Create API key",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/keys/{id}": {
      "delete": {
        "operationId": "delete-api-key",
        "parameters": [
          {
            "description": "Key UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Key UUID",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Delete API key",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "description": "Admin-only.",
        "operationId": "list-users",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List users",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "get-me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get current user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/users/me/activity": {
      "get": {
        "description": "Every SBOM ingested into a namespace you own, newest first. Scoped to the calling user: only resources you own, excluding public resources owned by others.",
        "operationId": "list-my-activity",
        "parameters": [
          {
            "description": "Maximum number of results per page",
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 20,
              "description": "Maximum number of results per page",
              "format": "int32",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a previous response's nextCursor; omit for the first page",
            "explode": false,
            "in": "query",
            "name": "cursor",
            "schema": {
              "description": "Opaque cursor from a previous response's nextCursor; omit for the first page",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListMyActivityOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List my recent activity",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/users/me/artifacts": {
      "get": {
        "description": "Scoped to the calling user: only resources you own, excluding public resources owned by others.",
        "operationId": "list-my-artifacts",
        "parameters": [
          {
            "description": "Maximum number of results per page",
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 20,
              "description": "Maximum number of results per page",
              "format": "int32",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a previous response's nextCursor; omit for the first page",
            "explode": false,
            "in": "query",
            "name": "cursor",
            "schema": {
              "description": "Opaque cursor from a previous response's nextCursor; omit for the first page",
              "type": "string"
            }
          },
          {
            "description": "Filter by artifact type",
            "explode": false,
            "in": "query",
            "name": "type",
            "schema": {
              "description": "Filter by artifact type",
              "type": "string"
            }
          },
          {
            "description": "Filter by artifact name",
            "explode": false,
            "in": "query",
            "name": "name",
            "schema": {
              "description": "Filter by artifact name",
              "type": "string"
            }
          },
          {
            "description": "Filter to artifacts with sufficiently enriched SBOMs; pass 'false' to include all (default: true)",
            "explode": false,
            "in": "query",
            "name": "sufficient",
            "schema": {
              "description": "Filter to artifacts with sufficiently enriched SBOMs; pass 'false' to include all (default: true)",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListArtifactsOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List my artifacts",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/users/me/namespaces": {
      "get": {
        "description": "Scoped to the calling user: only resources you own, excluding public resources owned by others.",
        "operationId": "list-my-namespaces",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListNamespacesOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List my namespaces",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/users/me/registries": {
      "get": {
        "description": "Scoped to the calling user: only resources you own, excluding public resources owned by others.",
        "operationId": "list-my-registries",
        "parameters": [
          {
            "description": "Maximum number of results per page",
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 20,
              "description": "Maximum number of results per page",
              "format": "int32",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Number of results to skip",
            "explode": false,
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "description": "Number of results to skip",
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRegistriesOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List my registries",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/users/me/sources": {
      "get": {
        "description": "Scoped to the calling user: only resources you own, excluding public resources owned by others.",
        "operationId": "list-my-sources",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSourcesOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List my sources",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/users/{id}/role": {
      "patch": {
        "description": "Admin-only.",
        "operationId": "update-user-role",
        "parameters": [
          {
            "description": "User UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "User UUID",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Update user role",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ]
}